requirements-telegram.txt
# Telegram Bot 名片管理系統依賴包

# Telegram Bot 核心
python-telegram-bot==20.7

# Web 框架
flask==3.0.0

# 圖片處理
pillow==10.1.0

# AI 識別
google-generativeai==0.3.2

# 資料庫操作
notion-client==2.2.1

# 環境變數管理
python-dotenv==1.0.0

# 網路請求
requests==2.31.0

# 日誌和工具
aiofiles==23.2.1